# api/v1alpha1/types.gen.go
// Package v1alpha1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package v1alpha1

import (
	"time"
)

// Error RFC 7807 problem details
type Error struct {
	Detail   *string `json:"detail,omitempty"`
	Instance *string `json:"instance,omitempty"`
	Status   *int32  `json:"status,omitempty"`
	Title    string  `json:"title"`
	Type     string  `json:"type"`
}

// Health Health status
type Health struct {
	Path  *string `json:"path,omitempty"`
	State string  `json:"state"`
	Type  *string `json:"type,omitempty"`
}

// Stack A 3-tier demo stack (web, app, db)
type Stack struct {
	CreateTime *time.Time `json:"create_time,omitempty"`

	// Id Unique stack ID
	Id *string `json:"id,omitempty"`

	// Name Human-readable stack name
	Name string `json:"name"`

	// Path Resource path
	Path       *string    `json:"path,omitempty"`
	UpdateTime *time.Time `json:"update_time,omitempty"`
}

// StackList defines model for StackList.
type StackList struct {
	NextPageToken *string  `json:"next_page_token,omitempty"`
	Stacks        *[]Stack `json:"stacks,omitempty"`
}

// ListStacksParams defines parameters for ListStacks.
type ListStacksParams struct {
	MaxPageSize *int32  `form:"max_page_size,omitempty" json:"max_page_size,omitempty"`
	PageToken   *string `form:"page_token,omitempty" json:"page_token,omitempty"`
}

// CreateStackJSONRequestBody defines body for CreateStack for application/json ContentType.
type CreateStackJSONRequestBody = Stack