# internal/api/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package server

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	. "github.com/dcm-project/3-tier-demo-service-provider/api/v1alpha1"
	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Health Check
	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// List Stacks
	// (GET /stacks)
	ListStacks(w http.ResponseWriter, r *http.Request, params ListStacksParams)
	// Create Stack
	// (POST /stacks)
	CreateStack(w http.ResponseWriter, r *http.Request)
	// Delete Stack
	// (DELETE /stacks/{stackId})
	DeleteStack(w http.ResponseWriter, r *http.Request, stackId string)
	// Get Stack
	// (GET /stacks/{stackId})
	GetStack(w http.ResponseWriter, r *http.Request, stackId string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Health Check
// (GET /health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List Stacks
// (GET /stacks)
func (_ Unimplemented) ListStacks(w http.ResponseWriter, r *http.Request, params ListStacksParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create Stack
// (POST /stacks)
func (_ Unimplemented) CreateStack(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete Stack
// (DELETE /stacks/{stackId})
func (_ Unimplemented) DeleteStack(w http.ResponseWriter, r *http.Request, stackId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get Stack
// (GET /stacks/{stackId})
func (_ Unimplemented) GetStack(w http.ResponseWriter, r *http.Request, stackId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListStacks operation middleware
func (siw *ServerInterfaceWrapper) ListStacks(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListStacksParams

	// ------------- Optional query parameter "max_page_size" -------------

	err = runtime.BindQueryParameter("form", true, false, "max_page_size", r.URL.Query(), &params.MaxPageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "max_page_size", Err: err})
		return
	}

	// ------------- Optional query parameter "page_token" -------------

	err = runtime.BindQueryParameter("form", true, false, "page_token", r.URL.Query(), &params.PageToken)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListStacks(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateStack operation middleware
func (siw *ServerInterfaceWrapper) CreateStack(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateStack(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteStack operation middleware
func (siw *ServerInterfaceWrapper) DeleteStack(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "stackId" -------------
	var stackId string

	err = runtime.BindStyledParameterWithOptions("simple", "stackId", chi.URLParam(r, "stackId"), &stackId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "stackId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteStack(w, r, stackId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStack operation middleware
func (siw *ServerInterfaceWrapper) GetStack(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "stackId" -------------
	var stackId string

	err = runtime.BindStyledParameterWithOptions("simple", "stackId", chi.URLParam(r, "stackId"), &stackId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "stackId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStack(w, r, stackId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/stacks", wrapper.ListStacks)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/stacks", wrapper.CreateStack)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/stacks/{stackId}", wrapper.DeleteStack)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/stacks/{stackId}", wrapper.GetStack)
	})

	return r
}

type GetHealthRequestObject struct {
}

type GetHealthResponseObject interface {
	VisitGetHealthResponse(w http.ResponseWriter) error
}

type GetHealth200JSONResponse Health

func (response GetHealth200JSONResponse) VisitGetHealthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListStacksRequestObject struct {
	Params ListStacksParams
}

type ListStacksResponseObject interface {
	VisitListStacksResponse(w http.ResponseWriter) error
}

type ListStacks200JSONResponse StackList

func (response ListStacks200JSONResponse) VisitListStacksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListStacks400ApplicationProblemPlusJSONResponse Error

func (response ListStacks400ApplicationProblemPlusJSONResponse) VisitListStacksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type ListStacks500ApplicationProblemPlusJSONResponse Error

func (response ListStacks500ApplicationProblemPlusJSONResponse) VisitListStacksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type CreateStackRequestObject struct {
	Body *CreateStackJSONRequestBody
}

type CreateStackResponseObject interface {
	VisitCreateStackResponse(w http.ResponseWriter) error
}

type CreateStack201JSONResponse Stack

func (response CreateStack201JSONResponse) VisitCreateStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateStack400ApplicationProblemPlusJSONResponse Error

func (response CreateStack400ApplicationProblemPlusJSONResponse) VisitCreateStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateStack409ApplicationProblemPlusJSONResponse Error

func (response CreateStack409ApplicationProblemPlusJSONResponse) VisitCreateStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type CreateStack500ApplicationProblemPlusJSONResponse Error

func (response CreateStack500ApplicationProblemPlusJSONResponse) VisitCreateStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type DeleteStackRequestObject struct {
	StackId string `json:"stackId"`
}

type DeleteStackResponseObject interface {
	VisitDeleteStackResponse(w http.ResponseWriter) error
}

type DeleteStack204Response struct {
}

func (response DeleteStack204Response) VisitDeleteStackResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteStack404ApplicationProblemPlusJSONResponse Error

func (response DeleteStack404ApplicationProblemPlusJSONResponse) VisitDeleteStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type DeleteStack500ApplicationProblemPlusJSONResponse Error

func (response DeleteStack500ApplicationProblemPlusJSONResponse) VisitDeleteStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetStackRequestObject struct {
	StackId string `json:"stackId"`
}

type GetStackResponseObject interface {
	VisitGetStackResponse(w http.ResponseWriter) error
}

type GetStack200JSONResponse Stack

func (response GetStack200JSONResponse) VisitGetStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetStack404ApplicationProblemPlusJSONResponse Error

func (response GetStack404ApplicationProblemPlusJSONResponse) VisitGetStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetStack500ApplicationProblemPlusJSONResponse Error

func (response GetStack500ApplicationProblemPlusJSONResponse) VisitGetStackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Health Check
	// (GET /health)
	GetHealth(ctx context.Context, request GetHealthRequestObject) (GetHealthResponseObject, error)
	// List Stacks
	// (GET /stacks)
	ListStacks(ctx context.Context, request ListStacksRequestObject) (ListStacksResponseObject, error)
	// Create Stack
	// (POST /stacks)
	CreateStack(ctx context.Context, request CreateStackRequestObject) (CreateStackResponseObject, error)
	// Delete Stack
	// (DELETE /stacks/{stackId})
	DeleteStack(ctx context.Context, request DeleteStackRequestObject) (DeleteStackResponseObject, error)
	// Get Stack
	// (GET /stacks/{stackId})
	GetStack(ctx context.Context, request GetStackRequestObject) (GetStackResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetHealth operation middleware
func (sh *strictHandler) GetHealth(w http.ResponseWriter, r *http.Request) {
	var request GetHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealth(ctx, request.(GetHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthResponseObject); ok {
		if err := validResponse.VisitGetHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListStacks operation middleware
func (sh *strictHandler) ListStacks(w http.ResponseWriter, r *http.Request, params ListStacksParams) {
	var request ListStacksRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListStacks(ctx, request.(ListStacksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListStacks")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListStacksResponseObject); ok {
		if err := validResponse.VisitListStacksResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateStack operation middleware
func (sh *strictHandler) CreateStack(w http.ResponseWriter, r *http.Request) {
	var request CreateStackRequestObject

	var body CreateStackJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateStack(ctx, request.(CreateStackRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateStack")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateStackResponseObject); ok {
		if err := validResponse.VisitCreateStackResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// DeleteStack operation middleware
func (sh *strictHandler) DeleteStack(w http.ResponseWriter, r *http.Request, stackId string) {
	var request DeleteStackRequestObject

	request.StackId = stackId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteStack(ctx, request.(DeleteStackRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteStack")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteStackResponseObject); ok {
		if err := validResponse.VisitDeleteStackResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetStack operation middleware
func (sh *strictHandler) GetStack(w http.ResponseWriter, r *http.Request, stackId string) {
	var request GetStackRequestObject

	request.StackId = stackId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetStack(ctx, request.(GetStackRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetStack")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetStackResponseObject); ok {
		if err := validResponse.VisitGetStackResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}